RDKB-60151: DscpCountPerInterval tr181 returns incorrect value
Reason for change: Whenever the DscpCountEnable is disabled and enabled, TX and RX takes the DscpCountTotal value. So, for the new client entries, I have made changes to assign the TX and RX to 0. After this the issue is not seen.

Test Procedure: Please refer the ticket

Risks: None

Priority: P1

Change-Id: I476e1dd0b27fff085444b79cf5ede468648f903e
Signed-off-by: umasankar.srinivasan@sky.uk